Build table mapping from snapshot job info
The upstream may rename the tables associated with the backup job.
Consequently, when the restore job is completed, the downstream table
won't be located when establishing the table mapping. This PR stores
the table name mapping from the snapshot job information, thus
preventing the need to read from the upstream.